test(split-pane): migrate
