+++ core/src/components/split-pane/test/basic/split-pane.e2e.ts-snapshots/split-pane-ios-ltr-Mobile-Chrome-linux.png
