+++ core/src/components/split-pane/test/basic/split-pane.e2e.ts-snapshots/split-pane-md-rtl-Mobile-Safari-linux.png
